투두 리스트 구현

+++ to-do-list-back-single-module/src/main/java/com/jiinkim/todolist/todo/controller/TodoController.java
@@ -3,9 +3,10 @@ package com.jiinkim.todolist.todo.controller;
 
 
 import com.jiinkim.todolist.common.dto.ApiResponse;
-import com.jiinkim.todolist.todo.application.service.dto.TodoGetResponse;
-import com.jiinkim.todolist.todo.application.service.dto.TodoInsertRequest;
-import com.jiinkim.todolist.todo.application.service.TodoService;
+import com.jiinkim.todolist.todo.dao.query.dto.TodoQueryDto;
+import com.jiinkim.todolist.todo.service.dto.TodoGetResponse;
+import com.jiinkim.todolist.todo.service.dto.TodoInsertRequest;
+import com.jiinkim.todolist.todo.service.TodoService;
 import com.jiinkim.todolist.todo.controller.dto.TodoServiceDtoConverter;
 import com.jiinkim.todolist.todo.controller.dto.TodoInsertClientRequest;
 import com.jiinkim.todolist.user.service.dto.UserDetailsImpl;
@@ -15,6 +16,9 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequiredArgsConstructor
@@ -33,8 +37,9 @@ public class TodoController {
     }
 
     @GetMapping
-    public ApiResponse<TodoGetResponse> getTodoList(@RequestParam("page")Long page
+    public ApiResponse<TodoGetResponse> getTodoList(@RequestParam("page")int page
         , @AuthenticationPrincipal UserDetailsImpl userDetails) {
+//        TodoListGetRequest dto = TodoServiceDtoConverter.of(page, userDetails.getUserId());
         return ApiResponse.success(HttpStatus.OK, todoService.getTodoList(page, userDetails.getUserId()));
     }
